% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders.R
\name{addToDo}
\alias{addToDo}
\title{Add ToDo item}
\usage{
addToDo(
  folder = NA,
  project = NA,
  entryDate = NA,
  dueDate = NA,
  scheduledDate = NA,
  reqTime = NA,
  description = NA
)
}
\arguments{
\item{folder}{Path to LabBookR folder}

\item{project}{The project name}
}
\value{
a list of projects
}
\description{
This functions adds a ToDo item to a project
}
