% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{checkLabBook}
\alias{checkLabBook}
\title{Check lab book}
\usage{
checkLabBook(labBook)
}
\value{
A matrix shoing the test results
}
\description{
This function checks, if the lab book is properly configured and prepared
}
