% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creaters.R
\name{createLabBook}
\alias{createLabBook}
\title{Create Complete LabBook}
\usage{
createLabBook(
  labBook = NULL,
  sortedByDate = TRUE,
  title = "My LabBook",
  author = "Daniel Fischer",
  output = "html+pdf"
)
}
\arguments{
\item{labBook}{Path to LabBookR folder}

\item{output}{Define output format, options: "html+pdf", "html", "pdf"}
}
\value{
A RMarkdown file
}
\description{
Create the lab book
}
