% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creaters.R
\name{createProjectReport}
\alias{createProjectReport}
\title{Create Project Report}
\usage{
createProjectReport(project = NULL, labBook = NULL)
}
\arguments{
\item{project}{Project name}

\item{labBook}{Path to LabBookR folder}
}
\value{
A RMarkdown dfile
}
\description{
Create a project report
}
