% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fixTime}
\alias{fixTime}
\title{Fix time values}
\usage{
fixTime(time)
}
\arguments{
\item{time}{Sting of time entry}
}
\value{
A hh:mm format string
}
\description{
This functions fixes inconsistent time entries
}
