% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getMyProjects}
\alias{getMyProjects}
\title{Show projects}
\usage{
getMyProjects(folder)
}
\arguments{
\item{folder}{Path to LabBookR folder}
}
\value{
a list of projects
}
\description{
This functions shows exixting project
}
