% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creaters.R
\name{createLabBook}
\alias{createLabBook}
\title{Create Complete LabBook}
\usage{
createLabBook(labBook = NULL, title = "My LabBook", author = "Daniel Fischer")
}
\arguments{
\item{labBook}{Path to LabBookR folder}
}
\value{
A RMarkdown file
}
\description{
Create the lab book
}
