% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creaters.R
\name{createNewProject}
\alias{createNewProject}
\title{Create a new project}
\usage{
createNewProject(title = NULL, labBook = NULL, author = NULL)
}
\arguments{
\item{title}{Name of the project}

\item{labBook}{Path to LabBookR folder}

\item{author}{Name of the project report author}
}
\value{
A blank project file
}
\description{
This function creates a new project
}
